[add] dealing with AI_IDs

+++ include/structs.h
@@ -642,7 +642,7 @@ typedef struct Unit {
     u16 _id;
     u16 title_id;
     i16 class;
-    i16 ai_id; /* Default AI for unit */
+    i32 ai_id; /* Default AI for unit */
 
     i8  mvt_type;
     i8  army;
@@ -720,7 +720,6 @@ typedef struct Unit {
     b32 mounted;
 
     s8 name;        /* TODO: get rid of it. Use id for global_unitNames */
-    // s8 title;       /* TODO: get rid of it. Use id for global_unitTitles */
 
     struct Computed_Stats computed_stats;   /* Computed from Unit_Stats */
